Initial import of ogmrip-0.12.3:
OGMRip is an application and a set of libraries for ripping and
encoding DVD into AVI, OGM, MP4, or Matroska files using a wide
variety of codecs. It relies on mplayer, mencoder, ogmtools,
mkvtoolnix, mp4box, oggenc, lame, and faac to perform its tasks.

The GUI features a clean HIG-compliant GNOME 2 interface and tries
to minimize as much as possible esoteric settings.

A CLI client is also available.

TODO:
Heavily dependent on Linux/FreeBSD features.
Might be possible to make it work there without much effort.

0.13.0 out.