Import gnubg-cvs-2009.07.21nb20090721 as wip/gnubg.
GNU Backgammon (gnubg) is for playing and analysing backgammon positions, games
and matches. It's based on a neural network. In the past twelve months it has
made enormous progress. It currently plays at about the level of a championship
flight tournament player. Depending on its parameters and its luck in recent
games, it rates from around 1900 to 2000 on FIBS, the First Internet Backgammon
Server -- at its strongest, it ranks in the top 5 of over 6000 rated players
there) and is gradually improving; it should be somewhat stronger than this when
released. Since almost all of the CPU time required during supervised training
is spent performing rollouts, and rollouts can easily be performed in parallel,
it is hoped that users will be able to pool rollout results and collectively
train it to a level stronger than any individual could obtain.