Added 389-ds-base-1.2.2 as wip/389-ds-base.
The enterprise-class Open Source LDAP server. It is hardened by
real-world use, is full-featured, supports multi-master replication, and
already handles many of the largest LDAP deployments in the world. The 389
Directory Server can be downloaded for free and set up in less than an hour
using the graphical console.

This package provides the Directory Server itself