Initial import of pcsx-df-1.10. Does not work at all, but compiles fine.
TODO:
Plugins need to be ported for non-Linux.
Dumps core when looking for plugins in /usr/lib.

DESCR:
PCSX-df is a forked version of the dead PCSX emulator, with a nicer
interface and several improvements to stability and functionality.

PCSX-df uses the PSEMU plugin interface to provide most functionality;
without them, you will not be able to use it to play games.

PCSX-df has a very capable Internal HLE BIOS that can run many
games without problems. It is recommended that you use it. However,
if you own a real Playstation, you may be able to use your own BIOS
image. PCSX-df will find it in ~/.pcsx/bios/ if you place it there.