Ispell is a fast screen-oriented spelling checker that shows you your errors in the context of the original file, and suggests possible corrections when it can figure them out.
This package does not contain anything by itself - it is a "meta-package" that
depends on other aspell packages. Its sole purpose is to require dependencies so
users can install this package only and have all the ispell stuff (or what is
set in PKG_OPTIONS.ispell-dictionaries) pulled in by the package dependency
mechanism.

This package is different from the other packages in pkgsrc/meta-pkgs. It's
possible to build everything by default, as the meta-pkgs currently function, or
PKG_OPTIONS.ispell-dictionaries can be set to pull in only certain dictionaries.
This is meant to save an admin time by just cd'ing to
pkgsrc/wip/ispell-dictionaries, running `bmake; bmake install' and going away to
do other important things like eating, drinking, and other forms of hedonism.

Currently, with meta-pkgs, it's all or nothing. Utilizing the options.mk
framework which exists in pkgsrc, it's not necessary to install any other
packages or write and maintain any additional configuration files to accomplish
this task, and at the same time not having to monitor and execute the build of
all the package required. This package also makes use of ispell dictionary in
pkgsrc-wip.