Import obmenu-1.0 as wip/obmenu.
Obmenu is a menu editor designed for openbox. It's easy to use, allowing to get
the most out of the powerful Openbox menu system, while hiding the xml layout
from the user.
It also (optionally) installs a bunch of dynamic menus (pipe menus), such as
Gnome menus or a quick-navigator.
You can also use the obxml module to easily write pipe menus of your own in
Python.