Import rope-0.9.2nb1 as wip/rope.
Rope is a python refactoring library.

A basic desciption about code refactoring found on wikipedia:

Code refactoring is the process of changing a computer program's
source code without modifying its external functional behavior
in order to improve some of the nonfunctional attributes of the
software.
Advantages include improved code readability and reduced complexity
to improve the maintainability of the source code, as well as a
more expressive internal architecture or object model to improve
extensibility.