Initial import of smw-1.8beta2:
Super Mario War is a Super Mario multiplayer game.  The goal is to
stomp as many other Marios as possible to win the game.  It's a
tribute to Nintendo and the game Mario War by Samuele Poletto.

The game uses artwork and sounds from Nintendo games. We hope that
this noncommercial fangame qualifies as fair use work. We just
wanted to create this game to show how much we adore Nintendo's
characters and games.

Package is unfinished due to compilation errors:
_src/gamemodes.cpp: In member function 'virtual void CGM_ShyGuyTag::think()':
_src/gamemodes.cpp:1259: warning: unused variable 'playwarning'
_src/player.h:500: warning: 'MO_SuperFireball' is already a friend of 'CPlayer'
_src/ai.cpp: In member function 'virtual void CPlayerAI::Think(COutputControl*)':
_src/ai.cpp:348: error: no match for 'operator=' in 'itr = ((CPlayerAI*)this)->CPlayerAI::attentionObjects.std::map<_Key, _Tp, _Compare, _Alloc>::erase [with _Key = int, _Tp = AttentionObject*, _Compare = std::less<int>, _Alloc = std::allocator<std::pair<const int, AttentionObject*> >](itr)'
/usr/include/g++/bits/stl_tree.h:152: note: candidates are: std::_Rb_tree_iterator<std::pair<const int, AttentionObject*> >& std::_Rb_tree_iterator<std::pair<const int, AttentionObject*> >::operator=(const std::_Rb_tree_iterator<std::pair<const int, AttentionObject*> >&)
gmake: *** [build/ai.o] Error 1