Import py26-repoze.tm2-1.0a5 as wip/py-repoze.tm2.
repoze.tm2 is a middleware which uses the ZODB transaction manager
to wrap a call to its pipeline children inside a transaction.  This
is a fork of the repoze.tm package which depends only on the transaction
package rather than the entirety of ZODB (for users who don't rely on
ZODB).