Import paraview-3.8.0 as wip/paraview.
ParaView is a powerful multi-platform data analysis and
scientific visualization application that can be used to
create 3D visualizations for quantitative or qualitative
exploration of data.  ParaView can be used either
interactively or for batch processing.  It can run on
clusters and supercomputers to analyze extremely large
datasets as well as on laptops for smaller volumes of data.