Import cl-unicode-0.1.1 as wip/cl-unicode.
CL-UNICODE is a library which provides Common Lisp implementations with
knowledge about Unicode characters including their name, their general
category, the scripts and blocks they belong to, their numerical value, and
several other properties. It also provides the ability to replace the standard
syntax for reading Lisp characters with one that is Unicode-aware and is used
to enhance CL-PPCRE with Unicode properties.

CL-UNICODE is based on Unicode 5.1.

The code comes with a BSD-style license so you can basically do with it
whatever you want.