Import libee-0.3.0 as wip/libee.
CEE is an upcoming standard used to describe network events in a number
of normalized formats. It's goal is to unify they currently many different
representations that exist in the industry.

The core idea of libee is to provide a small but hopefully convenient
API layer above the CEE standard. However, CEE is not finished. At the time
of this writing, CEE is under heavy development and even some of its core
data structures (like the data dictionary and taxonmy) have not been fully
specified.

More information on CEE can be found: http://cee.mitre.org/