Import hs-vty-4.7.0.0 as wip/hs-vty.
vty is terminal GUI library in the niche of ncurses. It is intended to
be easy to use, have no confusing corner cases, and good support for
common terminal types.

Included in the source distribution is a program
test/interactive_terminal_test.hs that demonstrates the various
features.

If your terminal is not behaving as expected the results of the
test/interactive_terminal_test.hs program should be sent to the Vty
maintainter to aid in debugging the issue.

Notable infelicities: Sometimes poor efficiency; Assumes UTF-8
character encoding support by the terminal.