Import hs-hxt-regex-xmlschema-9.0.0 as wip/hs-hxt-regex-xmlschema.
This library supports full W3C XML Schema regular expressions
inclusive all Unicode character sets and blocks. The complete grammar
can be found under <http://www.w3.org/TR/xmlschema11-2/#regexs>. It is
implemented by the technique of derivations of regular
expressions. The W3C syntax is extended to support not only union of
regular sets, but also intersection, set difference, exor. Matching of
subexpressions is also supported. The library can be used for
constricting lightweight scanners and tokenizers. It is a standalone
library, no external regex libraries are used. This package is a
substitute for the old regex-xmlschema package.