Import navi2ch-current-2.0.0nb20110814 as wip/navi2ch-current.
Navi2ch is a viewer program dedicated to chatting in 2ch.net - the biggest,
most famous, influential BBS in Japan.

Although this software is really well-written, it's worth noting that most
opinions/informations found in 2ch.net are next to senseless.  Thus, Good
NetBSD users are really encouraged not to devote themselves to 2ch! :-)