(1) patches/patch-ab patches/patch-ab has been modifying autogen/configure. Because of configuration scheme changed a little bit (upstream), autogen/configure is to be generated from configure.in. So patch target is changed.
(2) patch-ad (removed or renamed as patch-ab)
Used be modifying configure.in. But now (kind of) renamed as patch-ab.
(Older patch-ab is no longer necessary, and this patch-ad is renamed.)

(3) patches/patch-configure.in-2
Interim fix for 'wrong binary format problem' not finding crtn.o related,
which is introduced by following (upstream) commit.
---------
2011-06-29  Glenn Morris  <rgm@gnu.org>

        * configure.in: Try to test for the required crt*.o files.
---------

(4) TODO
  Above fix is just a revert of above change. This revert should not
be the right fix. So describe this situation

(5) PLIST
Adjust a long with upstream

(6) distinfo
Changes on patch-ab and adding patch-configure.in-2