Import jruby-1.6.4 as wip/jruby.
JRuby is a 100% Java implementation of the Ruby programming language.
It is Ruby for the JVM.

JRuby provides a complete set of core "builtin" classes and syntax
for the Ruby language, as well as most of the Ruby Standard Libraries.
The standard libraries are mostly Ruby's own complement of .rb
files, but a few that depend on C language-based extensions have
been reimplemented. Some are still missing, but we hope to implement
as many as is feasible.